updated x-ua-compatible meta tag and updated Meta Desc
Meta description of the page was updated from " " to "Despite all that Javascript is capable of, it is frequently employed. Here are some things you can accomplish without Javascript using HTML and CSS, which are surprisingly impressive on their own."

A missed opportunity for optimization exists if the meta description is absent and Your page will be more challenging for Google to understand. So This helps to describe the intent of this repo on SERPs for both crawlers and users.

X-UA-Compatible META tag was changed to content = "IE=edge from content = "ie=edge and it was moved to line 4 since You should position the X-UA-Compatible META tag as near to the HEAD of the page as you can if you're utilizing it.
Reference: https://stackoverflow.com/questions/6771258/what-does-meta-http-equiv-x-ua-compatible-content-ie-edge-do